api/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.10
django-filter==23.5
djangorestframework==3.14.0
idna==3.6
importlib-metadata==7.0.1
Markdown==3.5.2
pytz==2024.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
zipp==3.17.0
